Added HAPong: a rack middleware for lightning fast HAProxy pings.
HAProxy checks the health of your app with constant "ping" requests.  By default, this is an "OPTIONS /" request, which is often treated as a request to the default route by merb.  If your default route is not a static page or a simple response, these ping requests by HAProxy can slow down your server significantly.